langchain/data_analysis/pyproject.toml
[project]
name = "data-analysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepagents>=0.5.3",
    "langchain-daytona>=0.0.5",
    "langchain-ollama>=1.1.0",
    "python-dotenv>=1.2.2",
    "slack-sdk>=3.41.0",
]
